pallet-grandpa-asp: extract out mock file
