+++ frame/grandpa-accountable-safety/src/tests.rs
@@ -15,118 +15,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{self as pallet_grandpa_accountable_safety, check_commit_signatures};
-use frame_support::parameter_types;
+#![cfg(test)]
+
+use crate::check_commit_signatures;
+use crate::mock::*;
 use sp_core::H256;
-use sp_finality_grandpa::{
-	accountable_safety::{Query, QueryResponse},
-	Commit, RoundNumber, SetId,
-};
+use sp_finality_grandpa::accountable_safety::{Query, QueryResponse};
 use sp_keyring::Ed25519Keyring;
-use sp_runtime::{testing::Header, traits::IdentityLookup};
-
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-
-frame_support::construct_runtime!(
-	pub enum Test where
-		Block = Block,
-		NodeBlock = Block,
-		UncheckedExtrinsic = UncheckedExtrinsic,
-	{
-		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-		GrandpaAccountableSafety: pallet_grandpa_accountable_safety::{Pallet, Event},
-	}
-);
-
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub BlockWeights: frame_system::limits::BlockWeights =
-		frame_system::limits::BlockWeights::simple_max(1024);
-}
-
-impl frame_system::Config for Test {
-	type BaseCallFilter = ();
-	type BlockWeights = ();
-	type BlockLength = ();
-	type DbWeight = ();
-	type Origin = Origin;
-	type Index = u64;
-	type BlockNumber = u64;
-	type Call = Call;
-	type Hash = H256;
-	type Hashing = sp_runtime::traits::BlakeTwo256;
-	type AccountId = u64;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = Event;
-	type BlockHashCount = BlockHashCount;
-	type Version = ();
-	type PalletInfo = PalletInfo;
-	type AccountData = pallet_balances::AccountData<u128>;
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-	type SS58Prefix = ();
-	type OnSetCode = ();
-}
-
-parameter_types! {
-	pub const BlockTimeout: u64 = 10;
-}
-
-impl crate::Config for Test {
-	type Event = Event;
-	type BlockTimeout = BlockTimeout;
-}
-
-fn new_test_ext() -> sp_io::TestExternalities {
-	let storage = frame_system::GenesisConfig::default()
-		.build_storage::<Test>()
-		.unwrap();
-
-	storage.into()
-}
-
-fn create_precommits(
-	target_hash: H256,
-	target_number: u64,
-	round: RoundNumber,
-	set_id: SetId,
-	authorities: Vec<Ed25519Keyring>,
-) -> Vec<crate::SignedPrecommit<H256, RoundNumber>> {
-	let mut precommits = Vec::new();
-	for keyring in authorities {
-		let precommit = grandpa::Precommit {
-			target_hash,
-			target_number,
-		};
-		let msg = grandpa::Message::Precommit(precommit.clone());
-		let encoded = sp_finality_grandpa::localized_payload(round, set_id, &msg);
-		let signature = keyring.sign(&encoded[..]).into();
-		let signed_precommit = grandpa::SignedPrecommit {
-			precommit,
-			signature,
-			id: keyring.public().into(),
-		};
-		precommits.push(signed_precommit);
-	}
-	precommits
-}
-
-fn create_commit(
-	target_hash: H256,
-	target_number: u64,
-	round: RoundNumber,
-	set_id: SetId,
-	authorities: Vec<Ed25519Keyring>,
-) -> Commit<H256, RoundNumber> {
-	sp_finality_grandpa::Commit {
-		target_hash,
-		target_number,
-		precommits: create_precommits(target_hash, target_number, round, set_id, authorities),
-	}
-}
 
 #[test]
 fn verify_commit_signatures() {
@@ -137,7 +32,7 @@ fn verify_commit_signatures() {
 	];
 	let round = 42;
 	let set_id = 4;
-	let commit = create_commit(H256::random(), 5, round, set_id, authorities);
+	let commit = create_commit(authorities, H256::random(), 5, round, set_id);
 	assert!(check_commit_signatures(&(commit, round, set_id)));
 }
 
@@ -152,7 +47,7 @@ fn accountable_safety_setup_and_submit_reply() {
 		let alice = &authorities[0].public().into();
 		let round = 42;
 		let set_id = 4;
-		let commit = create_commit(H256::random(), 5, round, set_id, authorities.clone());
+		let commit = create_commit(authorities.clone(), H256::random(), 5, round, set_id);
 		let block_not_included = (commit.clone(), round.clone(), set_id);
 		let new_block = (commit.clone(), round, set_id);
 
@@ -189,7 +84,7 @@ fn accountable_safety_proceed_to_previous_round() {
 
 		let round = 42;
 		let set_id = 4;
-		let commit = create_commit(H256::random(), 5, round, set_id, authorities.clone());
+		let commit = create_commit(authorities.clone(), H256::random(), 5, round, set_id);
 		let block_not_included = (commit.clone(), round.clone(), set_id);
 		let new_block = (commit.clone(), round, set_id);
